Move ColorType definition and remove bytes support.
